React Rewrite - Import and adjust everything from preact version
Everything should be working now except network tab and events, although
that's not going to take a long time.

Finally - this fixes #4 - I couldn't get the useContext to work when
migrating from preact. Using zustard fixed it and provided a cleaner way
to get the state in non component functions, win-win situation.